Improved SCLYObject wrapper and MlvlWrapper::MlvlArea wrapper
## Added SclyObject::get_layer_idx()
Returns the layer index that the object belongs to.

## Added SclyObject::get_area_idx()
Returns the area index that the object belongs to.

## Added SclyObject::get_internal_idx()
Returns the internal index that the object of the current area.

## Added MlvlWrapper::MlvlArea::object_id_from_layer_name(layer_name, internal_idx)
Returns an object id from the layer name and the internal index of the object.

## Added MlvlWrapper::MlvlArea::object_id_from_layer_id(layer_name, internal_idx)
Returns an object id from the layer id and the internal index of the object.