Improvements and Bug Fixes to DLR doors (#122)
Fixes wrongful hit detection is some cases and incorrect scripting for
Research Core and Cargo Freight Lift to Deck Gamma.
Technically also Fixes #62 but it's not perfect, this chunk of code is
quite a mess and requires a rewrite but I'm not capable enough for it so
I worked with what I had, tried the best of my abilities on this and
it's good enough for me.


![GM8E01_2025-06-26_18-07-12](https://github.com/user-attachments/assets/a60ac4e1-a9c2-412d-9d4d-f57d0381854d)

![GM8E01_2025-06-26_18-13-22](https://github.com/user-attachments/assets/ab3b3d02-c535-4791-87f2-8839eb526951)

![GM8E01_2025-06-26_18-13-29](https://github.com/user-attachments/assets/278e51dc-3e51-478b-a45e-073cee2f976d)

![GM8E01_2025-06-26_18-13-37](https://github.com/user-attachments/assets/e8a85f0a-7164-4567-98ce-fe773298c6b2)

![GM8E01_2025-06-26_18-14-38](https://github.com/user-attachments/assets/1571012e-113a-4a05-9155-84fd17e0a73f)

![GM8E01_2025-06-26_18-15-54](https://github.com/user-attachments/assets/e53bd639-95e1-4ca3-89ac-a0a82614cd37)

![GM8E01_2025-06-26_18-17-34](https://github.com/user-attachments/assets/49233be3-b77b-47c0-a410-c5a6d0424ced)

![GM8E01_2025-06-26_18-18-55](https://github.com/user-attachments/assets/2da07bae-44ca-4cbd-b9ca-4fd50c526b6a)